Fix keepalive problems with the API server.
So it seems that the connection is just dropped without any more output after
doing a DELETE request followed by a bunch of other requests. I'm not sure if
the problem is just caused by DELETE, but observations showed so far.

We address this by checking for BadStatusLine exceptions in httplib and if one
happens, just disconnect, connect again and retry.