hetznerctl: Add option avoid running rescue shell.
This is especially useful if you want to activate the rescue system for several
servers and don't want to spawn a rescue shell for every server.

Especially because the normal behaviour is that the server is rebooted back into
the normal system after the rescue shell has been closed.

Now, the newly added -n option just prints out the passwords of the root user
for each server where the rescue shell has been activated _without_ rebooting
into the normal system afterwards.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>