server: Provide ability to create admin accounts.
As mentioned in the previous commit, this is implemented by using the
web *user* interface, which also means, that you only can create admin
accounts if you provide your real Robot credentials instead of the
separate webservice credentials that can be created in the Robot.

This somehow reminds me a bit about chroot() & companions where you had
to get root in order to restrict the process further. :-)

Also, the admin property is designed to be loaded lazily, so scraping
will only be used if you actually try to access the 'server.admin'
property.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>