Don't be specific about IPv4/IPv6.
This is somewhat blurry as the /ip/ and /subnet/ requests can contain
either IPv4 OR IPv6 addresses (not sure on /ip/ though) with no flag
that specifically distinguishes them.

So, in the future we might want to add our own detection routine to
decide whether an IP address is IPv4 or IPv6 (which should be fairly
easy).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>