server: Allow to fetch rDNS info for a Server.
This now introduces the ReverseDNSManager I wanted to avoid. But
iterating through all subnets of a particular server really is too
costly and we want to avoid that and keep the Robot API traffic low.

Also, this again relies on exception handling in business logic, which
is really bad, so let's fix this soon.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>