hetznerctl: Replace encoding errors by '?'.
It's not nice to rely on exception handling here, but we're doing this
only for hetznerctl, so it shouldn't be critical for the API itself.

This at least ensures that the encoding won't fail if the current locale
doesn't contain characters we want to encode.

A better way would be if we could change the default encoding error
handler, but unfortunately, we can't do this if we want to ensure
compatibility.

This is also noted at:

http://www.python.org/dev/peps/pep-3116/#unicode-encoding-decoding-issues

Signed-off-by: aszlig <aszlig@redmoonstudios.org>