robot: Avoid to use module-level loggers.
While this may be working for hetznerctl, it's really not a good idea to
do this, because we might want to set up loggers after we're done with
the imports even for hetznerctl in the future.

The main problem with this is that the default logger setup will be used
when the logging system is set up at some later point. We truly want to
avoid that and in addition, we now have proper logger names for the
robot classes as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>