hetznerctl: Add command to manage admin accounts.
The functionality for this was already in the API, but we were only
using it for NixOps so far. Now it's available using 'hetznerctl admin'.

Right now there isn't a robot API call for doing this, so we're scraping
the robot web interface. So it might be subject to being broken whenever
Hetzner is changing the Web UI.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>