tests: Convert util.addr tests into unit tests.
Unfortunately for retaining Python 3 compatibility, using doctests it is
quite pesky to maintain. So, I'm switching over to unit tests entirely
with a static test list.

The reason I'm using a static list of tests instead of autodiscovery is
that I want the tests to fail if the actual test files aren't there.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>