robot: Change pinned robot CA to Equifax.
Hetzner has recently changed the CA and so we need to chonge our
hardcoded root CA certificate here as well. This is probably a bit
unfortunate, because we really should have a way to either use the CA
file from OpenSSL or at least provide a way to specify a different one.

But as for version 0.7, this will at least work around it until Hetzner
changes CA again.

Thanks to @Davorak for reporting this in NixOS/nixops#227.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>