server: Fix setting admin account passwords.
The Hetzner robot web interface now includes a CSRF token for setting
the password for the admin account. For fetching the CSRF token, we need
to first request the HTML page containing the form, parse out the token
and submit it.

For parsing the token, we now have new class in hetzner.util.scraping,
which should handle that for us. I could have used regular expressions
for that, but in order to prevent further errors I went with a
HTMLParser instead.

Thanks to @rbvermaa for reporting and testing.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Reported-by: Rob Vermaas <rob.vermaas@gmail.com>
Tested-by: Rob Vermaas <rob.vermaas@gmail.com>