server: Fix creating admin accounts
Recent changes in the Robot not only make it unnecessary to create an
admin account with a Hetzner-provided random password and then set our
own password but also make it *necessary* to provide a password in the
first place while creating the admin account.

So we now first get the status of the admin page to get the CSRF token
and then just decide whether we need to update an existing account or
create a new one and use the relevant path.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Reported-by: Graham Christensen <graham@grahamc.com>
Closes: #18
(cherry picked from commit 3238d0afbfbb6a7a6e316d9b2b855ce40f68b025)