robot: Fix login for RobotWebInterface
The login for the Robot is no longer done via its own backend but
instead relies on a new site https://accounts.hetzner.com/ to do the
login and pass back an OAauth token to the Robot.

We probably should use something like the requests library
(http://python-requests.org/) for doing this, because the implementation
here turned out to be not very DRY.

Regardless of that, I'm still waiting for the day Hetzner is going to
implement managing Admin accounts into their Robot API so we can trash
all this scraping code for good.

I've tested creating, listing and removing of an admin account and it
works fine now, which means that just the login method has changed but
the Robot still uses the same API.

Signed-off-by: aszlig <aszlig@nix.build>
Fixes: #26