setup.py: Add discovery for doctests
We already have all the python modules listed in setup.py, so we can use
this to gather all the doctests, turn them into unit tests and add it to
the set of tests to run.

Right now we don't have any doctests, but the first one will be coming
soon.

Signed-off-by: aszlig <aszlig@nix.build>