Properly encode non-scalar values in API requests
Whenever we want to use arrays or even nested dictionaries in an API
request, we have to encode the data in a way PHP recognizes and there
currently doesn't seem to be a way to directly submit JSON data.

For example for passing authorized keys, the request needs to be encoded
like this:

  authorized_key%5B0%5D=FPR1&authorized_key%5B1%5D=FPR2

This corresponds to the following dictionary in Python:

  {'authorized_key': ['FPR1', 'FPR2']}

Right now, we do not handle escaping of keys in associative arrays, but
we can implement this later once we need to do something like this, eg.
when submitting firewall rules. For passing authorized keys however,
this is not needed.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @farcaller