Expose server power state and reset options
Some servers don't have the "sw" reset type anymore but now have a new
"power" reset type, which causes a power button press on the machine.

In addition, there is now an operating_status, which indicates whether
the server is running or powered off.

Both of these are now available via the server.reset API and I added
another is_running() helper which exposes the operating status as a
boolean.

The running status now is also available via "hetznerctl show".

Signed-off-by: aszlig <aszlig@nix.build>