setup.py: Require at least Python 3.7
Since we do have a branch for the 0.8.x version and it's about time we
really get rid of all the Python 2 related cruft in our code base,
especially since Python 2 is no longer officially supported since ages.

Since official support for Python 3.6 ends by the end of the year[1] and
we're making an immediate release of version 1.x, I'd say version 3.7 is
reasonable enough to expect on most systems.

[1]: https://devguide.python.org/#status-of-python-branches

Signed-off-by: aszlig <aszlig@nix.build>