Switch to basic Nix Flake
The implementation is pretty much equivalent to what we have so far in
default.nix and release.nix, but the advantage of using a Flake is that
we can pin a nixpkgs version and also provide Hydra with all the
required parameters so we no longer need to provide external settings.

This not only makes the tests more reproducible but also helps against
amnesia, eg. when you forget the correct function attrs for release.nix.

Signed-off-by: aszlig <aszlig@nix.build>