Properly name authorized_keys argument
We use authorized_keys in RescueSystem.activate() but pass it verbatim
via "authkey" in RescueSystem.observed_deactivate(). I consider this an
API inconsistency, so we really should use authorized_keyS everywhere.

I not only changed this in the API but also in the ctl command, which
now allows multiple uses of "-a" and thus implies that more than one key
is allowed.

Signed-off-by: aszlig <aszlig@nix.build>